+++ src/_incydr_sdk/file_events/client.py
@@ -21,7 +21,7 @@ class InvalidQueryException(IncydrException):
             "\nTip: Make sure you're specifying your filter fields in dot notation. "
             "\nFor example, filter by 'file.archiveId' to filter by the archiveId field within the file object.)"
         )
-        if 'problems' in exception.response.json().keys():
+        if "problems" in exception.response.json().keys():
             self.message += f"\nRaw problem data from the response: {exception.response.json()['problems']}"
         self.original_exception = exception
         super().__init__(self.message)